% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\docType{data}
\name{nba}
\alias{nba}
\title{Seasonal NBA stats}
\format{
A data frame with 24691 observations and 52 variables
}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/data}
}
\usage{
nba
}
\description{
Seasonal NBA stats
}
\keyword{datasets}
