% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_matr.R
\name{numeric_matr}
\alias{numeric_matr}
\title{Correlation Matrix}
\usage{
numeric_matr(year)
}
\arguments{
\item{year}{Year of interest}
}
\description{
Returns correlation matrix of numeric variables NBA set after filtering to given year.
}
\examples{
numeric_matr(2015)
}
\keyword{correlation}
\keyword{matrix}
