% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest player}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{Year of interest}
}
\description{
Returns oldest player during a given year. If multiple oldest players, returns the first one alphabetically by first name.
}
\examples{
oldest_player(2015)
"Andre Miller"
}
\keyword{oldest}
\keyword{player}
