% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_points.R
\name{player_points}
\alias{player_points}
\title{Player points}
\usage{
player_points(year, player = "Michael Jordan*")
}
\arguments{
\item{year}{Year of interest}

\item{player}{default=Michael Jordan}
}
\description{
Returns the points scored by a given player (Michael Jordan by default) in a given year.
}
\examples{
player_points(2015)
whatever the points are
}
\keyword{points}
